test:테스트 코드 디렉토리 재구성 및 CarsTest 분리
